Move src/sidebar/util/cfi.ts => src/shared/
Move these functions so they can be used by EPUB integrations in the
annotator code.
