+++ src/shared/test/cfi-test.js
